refactor: move empty calendar  modal file
