+++ src/actions/calendar-actions-fn.tsx
@@ -16,12 +16,12 @@ import { getFolderRequest } from '../soap/get-folder-request';
 import { getShareInfoRequest } from '../soap/get-share-info-request';
 import { CreateGroupModal } from './modals/create-group-modal';
 import { DeleteModal } from './modals/delete-modal';
+import { EmptyModal } from './modals/empty-modal';
 import { folderAction } from '../store/actions/calendar-actions';
 import { StoreProvider } from '../store/redux';
 import { ActionsClick } from '../types/actions';
 import { NewModal } from '../view/move/new-calendar-modal';
 import { EditModal } from '../view/sidebar/edit-modal/edit-modal';
-import { EmptyModal } from '../view/sidebar/empty-modal';
 import { ShareCalendarModal } from '../view/sidebar/share-calendar-modal';
 import { SharesInfoModal } from '../view/sidebar/shares-info-modal';
 import { SharesModal } from '../view/sidebar/shares-modal';
